1. remove cryptopp project use dotnet build crypt 2. convert all cryptopp integer to BigInteger

+++ src/utils/Mpd.Xml/Serialization/TypeData.cs
@@ -411,7 +411,7 @@ namespace Mpd.Xml.Serialization
 			"sizeof","while","lock","stackalloc","else","static","enum",
 			"namespace",
 			"object","bool","byte","float","uint","char","ulong","ushort",
-			"decimal","int","sbyte","short","double","long","string","void",
+			"BigInteger","int","sbyte","short","double","long","string","void",
 #if NET_2_0
 			"partial", "yield", "where"
 #endif
